Code AssetService and AssetAssignmentService
Also start coding mock repos and db for unit tests

+++ ITAssetTracking.Core/Interfaces/Repositories/IAssetAssignmentRepository.cs
@@ -7,7 +7,7 @@ public interface IAssetAssignmentRepository
     AssetAssignment? GetAssetAssignmentById(int assetAssignmentId);
     
     List<AssetAssignment> GetAllAssetAssignments(bool includeReturned);
-    List<AssetAssignment> GetAssetAssignmentsByAssetId(int assetId, bool includeReturned);
+    List<AssetAssignment> GetAssetAssignmentsByAssetId(long assetId, bool includeReturned);
     List<AssetAssignment> GetAssetAssignmentsByDepartmentId(int departmentId, bool includeReturned);
     List<AssetAssignment> GetAssetAssignmentsByEmployeeId(int employeeId, bool includeReturned);
     List<AssetAssignment> GetAssetAssignmentsInDateRange(DateTime startDate, DateTime endDate);
